Render laying player on server-side
Previously rest would simply update the players bed position metadata on
the client side to simulate the player laying in a bed. While this would
correctly put the player in a sleeping position on the players client
side, the pure existance of the bed position is not enough to render the
sleeping player to the other players on the server.

To ensure this, rest now also manually updates the players pose in the
data watcher and sends this to the players on the server. This ensures
that other clients do a manual pose update of the targetted player that
used the laying feature.