Check if Entity is Seat & Marked Seats with NBT
Before this commit there would nto be a method to check if a seat was a seat nor were seats marked in the cleanest of ways. This prevented plugins in KTN such as "Hospital" to check if the entities were seats or not.

As of this commit, PlayerSit interface contains a method to check if an entity is a seat as well as seats being marked with NBT tags instead of with a custom name. This will allow seats to be detectable by other plugins as well as the seats being tracked in a neat way.